+++ src/runtime/components/Link.vue
@@ -97,6 +97,13 @@ export type LinkPropsKeys = 'to' | 'href' | 'target' | 'rel' | 'noRel' | 'extern
 export interface LinkSlots {
   default(props: { active: boolean }): any
 }
+
+// from upstream NuxtLink
+interface NuxtLinkDefaultSlotProps {
+  rel: string | null
+  target: '_blank' | '_parent' | '_self' | '_top' | (string & {}) | null
+  isExternal: boolean
+}
 </script>
 
 <script setup lang="ts">
@@ -185,7 +192,7 @@ function resolveLinkClass({ route, isActive, isExactActive }: any) {
 
 <template>
   <NuxtLink
-    v-slot="{ href, navigate, route: linkRoute, rel, target, isExternal, isActive, isExactActive }"
+    v-slot="{ href, navigate, route: linkRoute, rel, target, isExternal, isActive, isExactActive, ...rest }"
     v-bind="nuxtLinkProps"
     :to="to"
     custom
@@ -200,9 +207,9 @@ function resolveLinkClass({ route, isActive, isExactActive }: any) {
           disabled,
           href,
           navigate,
-          rel,
-          target,
-          isExternal,
+          rel: (rest as NuxtLinkDefaultSlotProps).rel,
+          target: (rest as NuxtLinkDefaultSlotProps).target,
+          isExternal: (rest as NuxtLinkDefaultSlotProps).isExternal,
           active: isLinkActive({ route: linkRoute, isActive, isExactActive })
         }"
       />
@@ -217,9 +224,9 @@ function resolveLinkClass({ route, isActive, isExactActive }: any) {
         disabled,
         href,
         navigate,
-        rel,
-        target,
-        isExternal
+        rel: (rest as NuxtLinkDefaultSlotProps).rel,
+        target: (rest as NuxtLinkDefaultSlotProps).target,
+        isExternal: (rest as NuxtLinkDefaultSlotProps).isExternal
       }"
       :class="resolveLinkClass({ route: linkRoute, isActive, isExactActive })"
     >